Refactored gtr to separate module.
Final refactor of DNA substitution models into separate modules.
